+++ src/modules/Network.py
@@ -14,11 +14,11 @@ class ConnectedLayer:
         """
         self.inSize = inSize
         self.outSize = outSize
-        self.w = tf.clip_by_value(tf.Variable(tf.ones([self.inSize, self.outSize]), dtype=tf.float64), 0, float("inf"), dtype=tf.float64)
-        self.b = tf.Variable(tf.zeros(self.outSize), dtype=tf.float64)
+        self.w = tf.clip_by_value(tf.Variable(tf.ones([self.inSize, self.outSize]), dtype=tf.float32), 0, float("inf"), dtype=tf.float32)
+        self.b = tf.Variable(tf.zeros(self.outSize), dtype=tf.float32)
 
-        self.x = tf.placeholder(tf.float64, [None, self.inSize])
-        self.y = tf.placeholder(tf.float64, [None, self.outSize])
+        self.x = tf.placeholder(tf.float32, [None, self.inSize])
+        self.y = tf.placeholder(tf.float32, [None, self.outSize])
         
         self.z = tf.matmul(self.x, self.w) + self.b
